+++ backend/src/main/java/com/example/demo/global/init/InitStgData.java
@@ -6,7 +6,7 @@ import com.example.demo.domain.chat.service.ChatRoomService;
 import com.example.demo.domain.major.entity.Major;
 import com.example.demo.domain.post.entity.Post;
 import com.example.demo.domain.user.entity.User;
-import com.example.demo.domain.user.entity.UserStatus;
+import com.example.demo.domain.user.enums.UserStatus;
 import com.example.demo.domain.user.role.Role;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
